Support autocompletion with table aliases
Table names are only searched for between FROM and WHERE (if present) and only on the same level as the alias usage, so false positives should not appear.
It might not be implemented in the most performant way since I'm not familiar with token traversal in CM 6, but I hope it's not too bad. I only tested small queries, which are fine.